fn_8026E0E4 case 0: seqDoPrgChange midi param u8->u32 (TTYD DoPrgChange takes wider midi in r26) drops clrlwi+masked-rlwinm truncation at inline seam (98.531->98.854)

+++ src/main/audio/synth_seq_dispatch.c
@@ -210,7 +210,7 @@ static inline u16 seqHandleStream(SeqStream* stream)
     return stream->val;
 }
 
-static inline void seqDoPrgChange(SynthMidiState* seq, u8 prg, u8 midi)
+static inline void seqDoPrgChange(SynthMidiState* seq, u8 prg, u32 midi)
 {
     ((SynthMidiCtrlBlock*)lbl_803AF550)->midiCtrl[gSynthCurrentVoiceSlotIndex][midi] = 0xFFFF;
     if (midi != 9)
